extract DocumentSavePreparationCommand
