+++ celements-model/src/main/java/com/celements/store/part/CelHibernateStoreDocumentPart.java
@@ -1,13 +1,12 @@
 package com.celements.store.part;
 
-import static com.celements.model.util.References.*;
 import static com.google.common.base.MoreObjects.*;
+import static com.google.common.base.Preconditions.*;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Date;
 import java.util.Iterator;
-import java.util.UUID;
 
 import org.hibernate.FlushMode;
 import org.hibernate.HibernateException;
@@ -18,18 +17,13 @@ import org.hibernate.SessionFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.xwiki.model.EntityType;
-import org.xwiki.model.reference.ClassReference;
 import org.xwiki.model.reference.DocumentReference;
 import org.xwiki.model.reference.EntityReference;
-import org.xwiki.model.reference.WikiReference;
 
 import com.celements.model.object.xwiki.XWikiObjectEditor;
 import com.celements.model.object.xwiki.XWikiObjectFetcher;
 import com.celements.store.CelHibernateStore;
 import com.celements.store.id.CelementsIdComputer.IdComputationException;
-import com.google.common.base.Optional;
-import com.google.common.base.Preconditions;
-import com.google.common.base.Strings;
 import com.xpn.xwiki.XWikiContext;
 import com.xpn.xwiki.XWikiException;
 import com.xpn.xwiki.doc.XWikiAttachment;
@@ -37,7 +31,6 @@ import com.xpn.xwiki.doc.XWikiDocument;
 import com.xpn.xwiki.monitor.api.MonitorPlugin;
 import com.xpn.xwiki.objects.BaseObject;
 import com.xpn.xwiki.objects.classes.BaseClass;
-import com.xpn.xwiki.store.XWikiStoreInterface;
 import com.xpn.xwiki.util.Util;
 
 //TODO CELDEV-626 - CelHibernateStore refactoring
@@ -46,44 +39,19 @@ public class CelHibernateStoreDocumentPart {
   private static final Logger LOGGER = LoggerFactory.getLogger(CelHibernateStore.class);
 
   private final CelHibernateStore store;
+  private final DocumentSavePreparationCommand savePrepCmd;
 
   public CelHibernateStoreDocumentPart(CelHibernateStore store) {
-    this.store = Preconditions.checkNotNull(store);
+    this.store = checkNotNull(store);
+    this.savePrepCmd = new DocumentSavePreparationCommand(store);
   }
 
   public void saveXWikiDoc(XWikiDocument doc, XWikiContext context, boolean bTransaction)
       throws XWikiException {
     logXWikiDoc("saveXWikiDoc - start", doc);
     boolean commit = false;
-    MonitorPlugin monitor = Util.getMonitorPlugin(context);
     try {
-      // Start monitoring timer
-      if (monitor != null) {
-        monitor.startTimer("hibernate");
-      }
-
-      // initialise before beginning store transaction because of possible #loadXWikiDoc
-      XObjectPreparer objPreparer = new XObjectPreparer(doc, context);
-
-      doc.setStore(store);
-      // Make sure the database name is stored
-      doc.getDocumentReference().setWikiReference(new WikiReference(context.getDatabase()));
-
-      if (bTransaction) {
-        store.checkHibernate(context);
-        SessionFactory sfactory = store.injectCustomMappingsInSessionFactory(doc, context);
-        bTransaction = store.beginTransaction(sfactory, context);
-      }
-      Session session = store.getSession(context);
-      session.setFlushMode(FlushMode.COMMIT);
-
-      // These informations will allow to not look for attachments and objects on loading
-      doc.setElement(XWikiDocument.HAS_ATTACHMENTS, (doc.getAttachmentList().size() != 0));
-      doc.setElement(XWikiDocument.HAS_OBJECTS, getXObjectFetcher(doc).exists());
-
-      // Let's update the class XML since this is the new way to store it
-      // TODO If all the properties are removed, the old xml stays?
-      setBaseClass(doc, context);
+      Session session = savePrepCmd.execute(doc, bTransaction, context);
 
       if (doc.hasElement(XWikiDocument.HAS_ATTACHMENTS)) {
         store.saveAttachmentList(doc, context, false);
@@ -138,7 +106,6 @@ public class CelHibernateStoreDocumentPart {
         // session.saveOrUpdate(doc);
       }
 
-      objPreparer.execute();
       deleteAndSaveXObjects(doc, context);
 
       if (context.getWiki().hasBacklinks(context)) {
@@ -162,28 +129,9 @@ public class CelHibernateStoreDocumentPart {
         LOGGER.error("saveXWikiDoc - failed {} for {}", (commit ? "commit" : "rollback"),
             doc.getDocumentReference(), exc);
       }
-
-      // End monitoring timer
-      if (monitor != null) {
-        monitor.endTimer("hibernate");
-      }
     }
 
     logXWikiDoc("saveXWikiDoc - end", doc);
-  }
-
-  private void setBaseClass(XWikiDocument doc, XWikiContext context) {
-    BaseClass bclass = doc.getXClass();
-    if (bclass != null) {
-      bclass.setDocumentReference(doc.getDocumentReference());
-      if (bclass.getFieldList().size() > 0) {
-        doc.setXClassXML(bclass.toXMLString());
-      } else {
-        doc.setXClassXML("");
-      }
-      // Store this XWikiClass in the context in case of recursive usage of classes
-      context.addBaseClass(bclass);
-    }
   }
 
   private void deleteAndSaveXObjects(XWikiDocument doc, XWikiContext context)
@@ -197,50 +145,6 @@ public class CelHibernateStoreDocumentPart {
     for (BaseObject obj : getXObjectFetcher(doc).iter()) {
       store.saveXWikiCollection(obj, context, false);
     }
-  }
-
-  private class XObjectPreparer {
-
-    private final XWikiDocument doc;
-    private final XWikiDocument origDoc;
-
-    XObjectPreparer(XWikiDocument doc, XWikiContext context) throws XWikiException {
-      this.doc = doc;
-      this.origDoc = loadOriginalDocument(context);
-    }
-
-    private XWikiDocument loadOriginalDocument(XWikiContext context) throws XWikiException {
-      XWikiDocument dummyDoc = new XWikiDocument(cloneRef(doc.getDocumentReference(),
-          DocumentReference.class));
-      dummyDoc.setLanguage(doc.getLanguage());
-      if (!doc.isNew() && doc.hasElement(XWikiDocument.HAS_OBJECTS) && getPrimaryStore(
-          context).exists(dummyDoc, context)) {
-        return getPrimaryStore(context).loadXWikiDoc(dummyDoc, context);
-      }
-      return dummyDoc;
-    }
-
-    void execute() throws IdComputationException {
-      for (BaseObject obj : getXObjectFetcher(doc).iter()) {
-        obj.setDocumentReference(doc.getDocumentReference());
-        if (Strings.isNullOrEmpty(obj.getGuid())) {
-          obj.setGuid(UUID.randomUUID().toString());
-        }
-        if (!obj.hasValidId()) {
-          Optional<BaseObject> existingObj = XWikiObjectFetcher.on(origDoc).filter(
-              new ClassReference(obj.getXClassReference())).filter(obj.getNumber()).first();
-          if (existingObj.isPresent() && existingObj.get().hasValidId()) {
-            obj.setId(existingObj.get().getId(), existingObj.get().getIdVersion());
-            LOGGER.debug("saveXWikiDoc - obj [{}] already existed, keeping id", obj);
-          } else {
-            long nextId = store.getIdComputer().computeNextObjectId(doc);
-            obj.setId(nextId, store.getIdComputer().getIdVersion());
-            LOGGER.debug("saveXWikiDoc - obj [{}] is new, computed new id", obj);
-          }
-        }
-      }
-    }
-
   }
 
   public XWikiDocument loadXWikiDoc(XWikiDocument doc, XWikiContext context) throws XWikiException {
@@ -459,13 +363,6 @@ public class CelHibernateStoreDocumentPart {
 
   private XWikiObjectFetcher getXObjectFetcher(XWikiDocument doc) {
     return XWikiObjectEditor.on(doc).fetch();
-  }
-
-  /**
-   * @return the cache store if one is configured, else it's self referencing
-   */
-  private XWikiStoreInterface getPrimaryStore(XWikiContext context) {
-    return context.getWiki().getStore();
   }
 
   private void logXWikiDoc(String msg, XWikiDocument doc) {
